partial working version

+++ opal/datatype/opal_datatype.h
@@ -117,7 +117,7 @@ typedef void (*pack_type)(const char *d, const char *s);
 typedef int (*pack_partial_type)(char **dst, char *src, size_t *count, int *index, size_t *totdisp,
                                  size_t *disp, size_t *max_data );
 typedef size_t (*pack_opt_partial_type)(char **dst, char *src, size_t *count, 
-		                 int *elem_count, int *index, size_t *totdisp,
+		                 size_t *elem_count, int *index, size_t *totdisp,
                                  size_t *disp, size_t *datasize, size_t *max_data );
 /*
  * The datatype description.
